Temporarily hiding the collapse frame containing my experimental interface to interacting with linked segment transposition parameters. It's easily resurrectable.
I'm thinking that we could consider the transposition stuff on linked segments as a kind of "phase 2" of the linked segments work. There's enough stuff to be finished off with the interface to the basic linked segment functionality, and a few bugs to fix with it. Once the basics are in and solid I'll press on with the advanced features like transposition.

git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/linked_segments_ian@12205 cf6a83e7-f017-4305-b0b3-678bde24f230