Replace the bleak winter scene dog memorial with another bleak splash that's a little more enduring.  This one is the official Rosegarden rose again, but this time on a rainy, dreary day.  No musical instruments, no bright sunshine, just roses in the rain.  It seems like a more fitting splash to go out on, in case this turns out to be the last release.
It's hard to say when the last release will be, but I'm kind of thinking every
new release might be the last one now.  Probably not quite yet, but who knows at
this point.  Rosegarden is sort of at the point where you stick a mirror under
its nose from time to time, and watch for it to stop fogging before you get out
the burying shovel, but you want to dig out that shovel, and keep it near the
front of the tool shed, just in case today turns out to be the day.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12285 cf6a83e7-f017-4305-b0b3-678bde24f230