Partial fix for broken solo functionality in edit views.  The buttons were being created with checkable == false, so attempting to toggle them achieved nothing.
The balance of the fix will involve finding out the current solo state when a
new view is created, and updating the checked state of the buttons at that time,
to keep everything in sync.  I have no clue how to do that off the top of my
head, so we'll see if I can figure it out before I run out of time and/or
attention.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12293 cf6a83e7-f017-4305-b0b3-678bde24f230