Adding an "unlink" operation for changing a linked segment into a deep copy. In the UI I've followed the pattern laid down by the "Turn Repeats into Copies" menu option, adding a "Turn Links into Copies" option.
To implement the undo for this unlink operation, I need to guarantee that the SegmentLinker (the object responsible for coordinating a set of linked segments) is still alive at the time the undo is actioned. This is why the SegmentLinkToCopyCommand holds a copy of the original segment before unlinking it. It's a blunt but effective solution, at the cost of holding a copy of the segment in memory (the segments may be big, there may be a lot of them). Should be OK for the time being is my assumption for most real life use-cases.

The lighter alternative would be to have the SegmentLinker as a ref counted object and just hold a reference to it in the SegmentLinkToCopyCommand.

git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/linked_segments_ian@12423 cf6a83e7-f017-4305-b0b3-678bde24f230