Some interim solutions to problems with linked segments when they are resized from the left, and when they have cut/copy/paste range functions performed on them. We may decide to implement better or alternative behaviours, but at least these behaviours are somewhere to start.
Resizing from the left: 

When resizing a linked segment from the left, the link is broken and the resize is performed on a deep copy. The remaining links stay as they were. The alternative would be to honour the resize in the links also, and this is perfectly possible but will take me a bit more development time (it was relatively easy just to break the link for now).

Cut/Copy/Paste Range

Basically when doing any "range" type operation, I've broken the link and done the operation on a deep copy. For Copy Range I guess that's the right option. For Cut Range there might be an alternative of honouring the cut in all the links as well. For the Paste Range we could repeat the paste in all the links, although this might get quite messy!

Undo for any of these ops which break the link should restore the link again.

git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12500 cf6a83e7-f017-4305-b0b3-678bde24f230