Commit this framework for a new command I'm thinking about, so I have an incentive not to talk myself out of implementing it.
I still have some details to sort out, but the idea with this thing is that
setting things up for anacrusis is a tedious and unpleasant process that's
begging to be made simpler with this handy dandy new function.

The workflow I'm going to promote in the tutorial is to scan the sheet music to
see that you need anacrusis, figure out how long to make your segment, draw it
on neat lines with the automatic snap feature, and then use Create Anacrusis to
shove it precisely to the left while adding a bar 0 to your composition.

It's kind of one of those things where I decided to write a tutorial, but I
decided it was going to be too much of a pain in the ass to explain how to do
all this crap.  Instead, I'll work out a way for Rosegarden to do all this crap
for the user, and teach people how to use the new feature.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12658 cf6a83e7-f017-4305-b0b3-678bde24f230