This is the first commit for the Composition Notification Project.
The key change is the renaming of Composition::notifyTrackDeleted() to 
notifyTracksDeleted() and making it public.  The idea is that those who are 
modifying the Composition (DeleteTracksCommand::execute() and 
AddTracksCommand::unexecute()) know best when the modification is complete 
and the Composition is once again in a consistent state.  Therefore, they 
should be responsible for notifying Composition's observers of the completion 
of their modifications.  The rest of the code changes are in support of this 
new approach.

The final result is that TrackButtons can now add itself as a 
CompositionObserver and find out when tracks are deleted.  It can then update 
itself to stay in sync.  This has been tested by commenting out the call to 
m_trackButtons->slotUpdateTracks() in TrackEditor::paintEvent().  It appears 
to work fine.

Testing should include deleting tracks, undoing the addition of tracks, and
continued use of the software noting any problems.  

Feedback on this will help shape future changes.  I am particularly concerned 
about the notification responsibilities being shifted outside of Composition.
I think it's an ok move for a large system like rg.  Others might not.  
Dissenting opinions and discussion appreciated.

Detailed Change Breakdown

* Composition
  - Changed notifyTrackDeleted() to notifyTracksDeleted() and made it 
    public.  This is the key change.  
  - Removed call to notifyTrackDeleted() from detachTrack().  The 
    responsibility has been shifted outside of Composition to 
    AddTracksCommand and DeleteTracksCommand.
  - CompositionObserver::trackDeleted() changed to tracksDeleted().
  - These delete notification routines now take a list of tracks that
    have been deleted.  This should reduce unnecessary updates when
    multiple tracks are being deleted.
* AddTracksCommand
  - unexecute() now calls Composition::notifyTracksDeleted() once the
    added tracks have been deleted.
* DeleteTracksCommand
  - execute() now calls Composition::notifyTracksDeleted() once the
    tracks have been deleted.
* SequenceManager
  - trackDeleted() has been changed to tracksDeleted() and the code
    has been updated to handle the list of deleted tracks.
* TrackParameterBox
  - setDocument() now subscribes to Composition as an observer.  This
    was a bug.
  - trackDeleted() has been changed to tracksDeleted() and the code
    has been updated to handle a list of deleted tracks.
* TrackButtons
  - TrackButtons now derives from CompositionObserver and adds itself
    as an observer in its ctor.
  - tracksDeleted() has been added and it calls slotUpdateTracks() to
    make sure the track buttons are in sync with the Composition. 
* Miscellaneous changes
  - Commented out the unused Composition::deleteTrack().
  - Removed a noisy RG_DEBUG from CompositionModelImpl.
  - Some code reformatting.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12730 cf6a83e7-f017-4305-b0b3-678bde24f230