Fix drums in stormy riders:
1. Reassign track 9 to use instrument #10
2. Change instrument #10 from "fixed" to "auto" and check the [x] percussion box

It may be the latter change wasn't supposed to have been necessary, and I could
just be patching around a bug.

I'm not sure how I feel about that.  If users have to make minor tweaks like
this on old files, it really isn't the end of the world.  It took seconds to
achieve.  It would be better to avoid the need, if possible, though.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12901 cf6a83e7-f017-4305-b0b3-678bde24f230