Tim Munro had doubts about this patch, and it wasn't really intended to be applied as is so much as to illustrate a point.  Ted Felix agreed to investigate more deeply and work out some decision eventually whether Tim's solution is the right approach, or maybe there is a deeper issue at play, and Tim's approach is just scraping mold off the surface of a piece of bread.
It's all well and good for Ted to dig deeper, but I decided to just go ahead and
commit the patch after user Holger Marzen tried it (along with Tim's other
patch) and prounced the result "da fukin' best Rosegarden ever!"

When Ted gets a chance to do some of his famously methodical puzzle solving, he
may end up reverting this patch.  In the meantime, one user was really pleased
with it as is, so why not just commit it for now and run with it?

Here's a copy of the discussion for quick reference:

This concerns bug #3560849 reported recently (2012-08-23) by Ted Felix:

    1. Launch rg.
    2. Studio > Manage Synth Plugins.
    3. Configure synth plug-in #1 by selecting a plug-in from the dropdown.
    4. Click on "Controls" button for #1.  Controls dialog will appear.
    5. Close the controls dialog.
    6. Close the Manage Synth Plugins dialog.
    7. Studio > Manage Synth Plugins.
    8. Click on "Controls" button next to #1.
    9. Crash.

I think I might have accidentally fixed this a few months ago, while
tackling a similar bug in my own copy of Rosegarden:

    1. Launch rg.
    2. Studio > Audio Mixer.
    3. Click on any one of the plug-in buttons below the sliders.
    4. Close the plug-in dialog.
    5. Close the Audio Mixer dialog.
    6. Studio > Audio Mixer.
    7. Click on the same plug-in button as before.
    8. Crash.

The problem, it seems, is in
"src/gui/application/RosegardenMainWindow.cpp."  The conditional
statement at line 7551 is apparently not doing its job, so (I'm
guessing) line 7586 can refer to a parent that probably no longer
exists.  I don't know how to fix this right, but my hack involves hard
coding line 7586 to refer directly to "this" rather than to "parent."

I occupy a small corner of the Rosegarden world, using only the
notation editor, along with DSSI-softsynth and LADSPA plugins.
Anything that I do to fix my own copy of Rosegarden could easily damage
some other aspect of the program, and I would never know it, because I
never go there.  My understanding of Rosegarden and C++ is limited.

That's a fancy way of saying that I don't mean for the patch that I've
attached to be applied as is.  I've included it only because it shows
exactly what I've done to make the problem go away.  It's one of about
a dozen patches that I routinely apply to my own copy of Rosegarden
(svn_12984 currently) to make life easier.

The last part of the patch concerning
"src/gui/dialogs/AudioPluginDialog.cpp" is not directly related to this
problem but does involve the LADSPA plugins.  It solves the highly
annoying problem of LADSPA-plugin controls initially displaying some
screwy default settings, instead of showing the settings previously
saved in the .rg file.  It's the same problem that Holger Marzen has
mentioned a couple of times.

Tim 



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13103 cf6a83e7-f017-4305-b0b3-678bde24f230