Fix #1361
It looks like this was a purely superficial bug that caused the display to
behave strangely without actually affecting the real setting of any affected
parameters.  This bug was courtesy of none other than yours truly, and I have no
idea WTF I was thinking.  It looks like I was just skimming API docs in a big
hurry, and I grabbed something that looked superficially correct, and rewrote
several lines as utter garbage.

I didn't study the mechanics exhaustively, and basically just looked for a
better way to rewrite the old Qt 3 code that was still helpfully left in
comments.  This looks right, and it passes a cursory test.

Please thrash this and let's be sure it's really working now.  In particular, I
didn't even look at the track parameter box before instituting the same sort of
changes over there, and I'm not even sure it was broken before.  It most likely
was, but this was just a blind guess.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13118 cf6a83e7-f017-4305-b0b3-678bde24f230