New "short label" for track, to be used as alternate staff name in LilyPond. I'm just getting something in the can here, and problems remain.
BUGS:

The style is totally mangled

If you're just printing out one part from a score, you get the label repeating
on every single line:
  

  Trumpet
  Tpt.
  Tpt.
  Tpt.

The short label makes great sense for a large score, but I think for 1-2 parts
the short label should be omitted.

Thinking about where to draw the line, I think just make a damn toggle and be
done with it.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13179 cf6a83e7-f017-4305-b0b3-678bde24f230