Add setting for UI update interval.
For performance testing as with the previous two commits.

This will have an effect on the Windows port which I believe has the
update interval set to something other than 50.  That number has
moved up to the settings call as the default:

int updateUITime = settings.value("Update_UI_Time", 50).toInt();


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13255 cf6a83e7-f017-4305-b0b3-678bde24f230