Fix subtle bug in track divider drawing.
The darker lines on the outside of the track dividers would not be drawn
if you scrolled vertically by single pixels.  Problem was that the
divider positions were being obtained using too narrow of a clip rect.

Changed the "p" argument to CompositionView::drawArea() to the more
helpful "segmentLayerPainter".

Various other renaming and code simplification.

Moved a number of public slots to protected.

Removed unnecessary use of "mutable".


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13292 cf6a83e7-f017-4305-b0b3-678bde24f230