Add new NotationPreview class.
This is a work in progress.  Committing it for a backup.  I'm trying to
pull the notation preview concept out of CompositionModelImpl.  This
is one of the areas that uses excessive CPU time.  Isolating it will
make it easier to swap it out with a more efficient implementation.

At this point, NotationPreview is unused code, so it should have no 
effect on the system.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13346 cf6a83e7-f017-4305-b0b3-678bde24f230