Add new segmenttool.rc.
Trying to fix some Qt warnings related to a misplaced menu
in rosegardenmainwindow.rc.  "make qrc" requires that files
be in version control.  So, here it is "make qrc".

BTW, there were several files that weren't in the qrc that
is in version control.  I'll get to them once this is
finished.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13376 cf6a83e7-f017-4305-b0b3-678bde24f230