PHASE 1 of n:  Remove the deprecated selection event filter from the event list editor.  I never liked having this here anyway, and only put it here because I caved to pressure from an influence long since gone.
The selection event filter is officially DEPRECATED.  As of this commit, I am
crossing the Rubicon, and exploring undiscovered country.  The selection event
filter will remain in place in the matrix and notation editors until its
replacement is fully implemented and debugged, which will probably be a release
cycle or two after the new search functionality goes into effect.

Instead of thinning a selection with a filter by subtracting, selections will be
built by adding using a hugely more configurable filter.  This will probably
take the form of Edit -> Search... and Edit -> Search Range... and related
actions.  The plan is still coming together, but I see it clearly enough to step
across the Rubicon and start heading that way.

It will aim to implement the nearly endless array of possibilities in stages, in
order of priority, and the new filter will let you zero in on very specific
things, like all the quarter notes with longer performance durations that really
want to be half notes in a split out voice, for instance.

If we don't get anything done about the tuplet problem before then, it might
could pick all the notes that are tuplets but not tuplets, so at least you could
correct them manually much more easily and conveniently.

And so it begins.  This is going to be a slog!


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13443 cf6a83e7-f017-4305-b0b3-678bde24f230