Comment out Q_ASSERT, as it is plainly not safe to use in production code.
I did a test in a clean bash environment (no extra environment variables)  with
a clean release build, and Rosegarden crashed on this assert.  I added
-DQT_NO_DEBUG_OUTPUT and -dQT_NO_WARNING_OUTPUT to configure.ac, did another
clean release build, verified that the flags were being passed to the comiler,
and yet Rosegarden continued to crash on this assert.  According to the Qt API
docs that should not be, but it is.  Plainly it is not safe to use Q_ASSERT in
production code.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13478 cf6a83e7-f017-4305-b0b3-678bde24f230