SelectDialog: make all the use_all buttons work.  Woof.  The guy who dreamed up this interface is a SADIST.
Now that the row and column buttons are working, it highlights some additional
problems with coordination and checked states.  I have more thinking to do, but
I'm liking how the concept is working out so far.  It's outside the box, that's
for sure.  It's too bad it would take almost as much work to prototype something
like this out in Designer as it does to just write the code the hard way.  It's
a lot of work to prove a concept from the back of a napkin!

Oh, thanks to Ted Felix for the Qt fail on warnings idea.  I caught a dozen bugs
that would have been insidious and tricky to discover otherwise, because staring
at this code makes your eyes cross and you go halfway blind.  RestDotted,
DottedRest, is it done yet?  Please?  Can't it be over soon?

(This monster has swallowed something like 12 hours so far, and I am just
scratching the surface of what this dialog is going to do eventually.  Ouch.)


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13502 cf6a83e7-f017-4305-b0b3-678bde24f230