SelectDialog work.  Make non-memory CheckButton default to being checked.
This seems to make everything behave a little more like I'd expect, but it
doesn't really address the problem of the arrow buttons getting out of sync with
the row or column they affect.  If everything in that row/column is turned off,
but the arrow button is checked, you have to press it twice to turn everything
on.

The alternative to living with this is pretty mind boggling, and I may just live
with this and spare myself a lot of work.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13503 cf6a83e7-f017-4305-b0b3-678bde24f230