Hack CheckButton with a local stylesheet (does not respect m_Thorn)
I figured out a way to hack the style after all.  Non-memory CheckButton widgets
do not paint themselves differently when they are checked, so there is no way to
tell.  Sometimes you don't get the result you're trying for on the first pull,
but it's pretty obvious you keep pushing it and you get it next time.  You no
longer end up with stray checked buttons on empty rows, or unchecked buttons on
full rows.  I can live with that.

I officially consider this crazy mess good enough for government work, and I'll
be moving on after this.  I'm maybe 25% of the way through writing the GUI, if
I'm lucky.  Woof.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@13504 cf6a83e7-f017-4305-b0b3-678bde24f230